clean_rename: Rename files and references
