+++ CMakeLists.txt
@@ -1301,10 +1301,10 @@ set(CLAIM_PLUGIN_FILES
 )
 
 set(CLAIM_WINDOWS_FILES
-        src/claim/netdata_claim.c
-        src/claim/netdata_claim.h
-        src/claim/netdata_claim_window.c
-        src/claim/netdata_claim_window.h
+        src/claim/main.c
+        src/claim/main.h
+        src/claim/ui.c
+        src/claim/ui.h
 )
 
 set(ACLK_ALWAYS_BUILD
@@ -2262,8 +2262,8 @@ add_executable(netdata
 )
 
 if(OS_WINDOWS)
-        add_executable(netdata_claim ${CLAIM_WINDOWS_FILES} ${NETDATA_CLAIM_RES_FILES})
-        target_link_libraries(netdata_claim shell32;gdi32;msftedit)
+        add_executable(NetdataClaim ${CLAIM_WINDOWS_FILES} ${NETDATA_CLAIM_RES_FILES})
+        target_link_libraries(NetdataClaim shell32;gdi32;msftedit)
 endif()
 
 target_compile_definitions(netdata PRIVATE
@@ -2456,7 +2456,7 @@ if(NOT OS_WINDOWS)
                 DESTINATION "${BINDIR}")
 else()
         install(PROGRAMS
-                ${CMAKE_BINARY_DIR}/netdata_claim.exe
+                ${CMAKE_BINARY_DIR}/NetdataClaim.exe
                 COMPONENT netdata
                 DESTINATION "${BINDIR}")
 endif()
